Hardcode all resource URLs.
This means we are no longer using Bower for dependency management, but instead
as something like a wrapper for wget or curl.

One effect of this seems to be for bower to fallback to `index.js` as a default
name for a main script.

Somehow this has made mocha unhappy with our regex approach to ignoring the
jquery global, so it is switched to a globbing approach.